# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GraphDB API",
    "description": "GraphDB network service layer (HTTP/gRPC)",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/sessions": {
      "get": {
        "tags": [
          "Session"
        ],
        "operationId": "get_v1_sessions",
        "responses": {
          "200": {
            "description": "Session list",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal error"
          }
        }
      },
      "post": {
        "tags": [
          "Session"
        ],
        "operationId": "post_v1_sessions",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Session created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error"
          }
        }
      }
    },
    "/v1/sessions/{id}": {
      "get": {
        "tags": [
          "Session"
        ],
        "operationId": "get_session",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Session id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session details",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal error"
          }
        }
      },
      "delete": {
        "tags": [
          "Session"
        ],
        "operationId": "delete_session",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Session id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Session deleted"
          },
          "500": {
            "description": "Internal error"
          }
        }
      }
    },
...
    "/v1/statistics/sessions/{id}": {
      "get": {
        "tags": [
          "Statistics"
        ],
        "summary": "Obtaining session statistics",
        "operationId": "session",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Session id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session statistics",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateSessionRequest": {
        "type": "object",
        "description": "Create session request",
        "required": [
          "username",
          "client_ip"
        ],
        "properties": {
          "client_ip": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        }
      },
...
      "SessionResponse": {
        "type": "object",
        "description": "Session response",
        "required": [
          "session_id",
          "username"
        ],
        "properties": {
          "created_at": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "session_id": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}